requirements.txt
Django
dramatiq
requests
python-whois